[ANDWL-162] Add ConnectWalletActivity
