+++ app/src/main/java/wannabit/io/cosmostaion/activities/MainActivity.java
@@ -1,7 +1,6 @@
 package wannabit.io.cosmostaion.activities;
 
 import static wannabit.io.cosmostaion.base.BaseChain.BNB_MAIN;
-import static wannabit.io.cosmostaion.base.BaseChain.CRESCENT_TEST;
 import static wannabit.io.cosmostaion.base.BaseChain.KAVA_MAIN;
 import static wannabit.io.cosmostaion.base.BaseChain.OKEX_MAIN;
 import static wannabit.io.cosmostaion.base.BaseChain.SIF_MAIN;
@@ -42,11 +41,9 @@ import java.util.ArrayList;
 
 import desmos.profiles.v1beta1.ModelsProfile;
 import wannabit.io.cosmostaion.R;
-import wannabit.io.cosmostaion.activities.chains.cresent.CresentConnectActivity;
 import wannabit.io.cosmostaion.activities.chains.desmos.ProfileActivity;
 import wannabit.io.cosmostaion.activities.chains.desmos.ProfileDetailActivity;
 import wannabit.io.cosmostaion.activities.chains.kava.ClaimIncentiveActivity;
-import wannabit.io.cosmostaion.activities.chains.kava.KavaWalletConnectActivity;
 import wannabit.io.cosmostaion.activities.chains.sif.SifIncentiveActivity;
 import wannabit.io.cosmostaion.base.BaseActivity;
 import wannabit.io.cosmostaion.base.BaseChain;
@@ -427,10 +424,8 @@ public class MainActivity extends BaseActivity implements FetchCallBack {
             Intent wIntent = null;
             if (mBaseChain.equals(BNB_MAIN)) {
                 wIntent = new Intent(this, WalletConnectActivity.class);
-            } else if (mBaseChain.equals(KAVA_MAIN)) {
-                wIntent = new Intent(this, KavaWalletConnectActivity.class);
-            } else if (mBaseChain.equals(CRESCENT_TEST)) {
-                wIntent = new Intent(this, CresentConnectActivity.class);
+            } else {
+                wIntent = new Intent(this, ConnectWalletActivity.class);
             }
             wIntent.putExtra("wcUrl", data.getStringExtra("wcUrl"));
             startActivity(wIntent);
